search_engine_with_tools_and_agents/requirements.txt
ipykernel
python-dotenv
langchain
langchain_community
langchain_core
langchain_classic
streamlit
openai
langchain_openai
huggingface_hub
langchain_groq
langchain_ollama
pypdf
langchain_huggingface
sentence-transformers
faiss-cpu
python-dotenv
langchain_chroma
arxiv
wikipedia
ddgs
